chore(gitignore): Ignoring eslintrc
For those in the world that use eslint, using standard wholesale may not be an
option. But there exist alternatives for using eslint while complying with
standard. Instead of adding feross/eslint-config-standard as an npm module, I'm
grabbing the config of that project and using it locally. Hence the ignore.