test(url): Testing for hoodie.url property
Testing for a url property on the constructed hoodie object. A base url my be
provided as an option and the constructor will append `/hoodie` to it as the url
property. If a url is not provided, then the url will be defaulted to
`window.location.origin`. To test this I created an object on global called
`window` that had a simulated `location.origin`.