feat: hoodie.url getter
Adding a url property to hoodie api as a getter. The url is passed in as an
option and stored in the state. If no url is provided in the options, then the
`state.url` will default to window.location.origin. The getter will return the
`state.url + /hoodie`. I'm also piping the url through to Account, Store, and
Task in the constructor.

closes #13