feat: hoodie.plugin
Adding a method to add 'plugins' to hoodie object. Plugins can be passed as a
function (a callback that will recieve hoodie as a parameter) or an object (the
key will be the method name, and the value is the function). Plugin calls can be
chained to add multiple plugins. See README.md for examples.

closes #2