test: removeAll() rejects with error if no docs in scope
Calling `hoodie.store(scopeName).removeAll()` when there are no documents with type: scopeName results in an uncaught error `Items not found`.  This is due to the fact that in `removeAll()` we call `remove()` even if the list of returned objects is empty.  `remove()` eventually calls `find()` which then throws an `Items not found` error.

Instead, `hoodie.store(scopeName).removeAll()` should just return an empty list.  This test will pass once the fix is in place.